oem: update firefly uvcc

+++ oem/oem_firefly_uvcc/aicamera.sh
@@ -2,6 +2,44 @@
 #
 
 TRY_CNT=0
+get_pid(){
+   ps -A | grep "$1" | awk '{print $1}'
+}
+wait_process_killed(){
+   if [ "$2" = "" ]; then return; fi
+   while [ "$(get_pid $1)" = "$2" ]
+   do
+     sleep 0.1
+   done
+}
+
+check_uvc_suspend()
+{
+  if [ -e /tmp/uvc_goto_suspend ];then
+     echo "uvc go to suspend now"
+     ispserver_pid=$(get_pid ispserver)
+     aiserver_pid=$(get_pid aiserver)
+     killall ispserver
+     killall aiserver
+     wait_process_killed ispserver ${ispserver_pid}
+     wait_process_killed aiserver ${aiserver_pid}
+     CNT=0
+     while [ "$CNT" -gt 20 ]
+     do
+       if [ -e /tmp/uvc_goto_suspend ];then
+          sleep 0.1
+          let CNT=CNT+1
+       else
+          CNT=100
+       fi
+     done
+     if [ -e /tmp/uvc_goto_suspend ];then
+       rm /tmp/uvc_goto_suspend -rf
+       echo mem > /sys/power/state
+     fi
+  fi
+}
+
 check_uvc_buffer()
 {
   if [ "$TRY_CNT" -gt 0 ];then
@@ -22,14 +60,26 @@ check_uvc_buffer()
 }
 check_alive()
 {
+  if [[ ! -f "/oem/usr/bin/$1"  && ! -f "/usr/bin/$1" ]]; then
+   return 1
+  fi
   PID=`busybox ps |grep $1 |grep -v grep | wc -l`
   if [ $PID -le 0 ];then
      if [ "$1"x == "uvc_app"x ];then
        echo " uvc app die ,restart it and usb reprobe !!!"
+       killall adbd
+       killall uac_app &
        sleep 1
+       killall -9 adbd
+       killall -9 uac_app
        rm -rf /sys/kernel/config/usb_gadget/rockchip/configs/b.1/f*
-       echo ffd00000.dwc3  > /sys/bus/platform/drivers/dwc3/unbind
-       echo ffd00000.dwc3  > /sys/bus/platform/drivers/dwc3/bind
+       echo none > /sys/kernel/config/usb_gadget/rockchip/UDC
+       rmdir /sys/kernel/config/usb_gadget/rockchip/functions/rndis.gs0
+       rmdir /sys/kernel/config/usb_gadget/rockchip/functions/ffs.adb
+       rmdir /sys/kernel/config/usb_gadget/rockchip/functions/uac*
+       UDC=`ls /sys/class/udc/| awk '{print $1}'`
+       echo $UDC  > /sys/bus/platform/drivers/dwc3/unbind
+       echo $UDC  > /sys/bus/platform/drivers/dwc3/bind
        /oem/usb_config.sh rndis off #disable adb
        usb_irq_set
        uvc_app &
@@ -41,6 +91,8 @@ check_alive()
             echo "aiserver is die,tell uvc to recovery"
             killall -3 uvc_app
             aiserver &
+            sleep .5
+            killall -10 smart_display_service
          else
             $1 &
          fi
@@ -70,17 +122,22 @@ usb_irq_set()
 dbserver &
 ispserver -n &
 stop_unused_daemon
+#uac_app &
 /oem/usb_config.sh rndis
 usb_irq_set
 uvc_app &
 aiserver &
+sleep .5
+smart_display_service &
 while true
 do
   check_alive dbserver
   check_alive ispserver
   check_alive uvc_app
+#  check_alive uac_app
   check_alive aiserver
-  check_uvc_buffer
+#  check_uvc_buffer
+#  check_uvc_suspend
   sleep 2
   check_alive smart_display_service
 done
